- move image manipulation to BitmapUtils - show progress of scan by number of points collected
